# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: sign-verify-2 API
  version: 1.0.0
  description: REST service for electronic signature verification, extraction, profile and key management.
servers:
  - url: /
tags:
  - name: ApiKeys
  - name: Profiles
  - name: Verifications
  - name: Extractions
  - name: Tsl
  - name: Audit
  - name: Health

security:
  - ApiKeyAuth: []
  - OAuth2Bearer: []

paths:
  /actuator/health:
    get:
      tags: [Health]
      operationId: health
      security: []
      responses:
        '200': { description: ok }

  /api/v1/api-keys:
    get:
      tags: [ApiKeys]
      operationId: listApiKeys
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyPage' }
    post:
      tags: [ApiKeys]
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiKeyCreateRequest' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyCreatedResponse' }

  /api/v1/api-keys/{id}:
    delete:
      tags: [ApiKeys]
      operationId: deleteApiKey
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: deleted }
    patch:
      tags: [ApiKeys]
      operationId: patchApiKey
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiKeyPatchRequest' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyView' }

  /api/v1/profiles:
    get:
      tags: [Profiles]
      operationId: listProfiles
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfilePage' }
    post:
      tags: [Profiles]
      operationId: createProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileCreateRequest' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileView' }
  /api/v1/profiles/{id}:
    get:
      tags: [Profiles]
      operationId: getProfile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileView' }
    put:
      tags: [Profiles]
      operationId: updateProfile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdateRequest' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileView' }
    delete:
      tags: [Profiles]
      operationId: deleteProfile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: deleted }
  /api/v1/profiles/{id}/default:
    post:
      tags: [Profiles]
      operationId: setDefaultProfile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileView' }

  /api/v1/verifications:
    post:
      tags: [Verifications]
      operationId: verifySync
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                metadata:
                  type: string
                  description: "JSON con profileId?, profileOverrides?, reports[]?"
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationResponse' }

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    OAuth2Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, format: int32 }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }

    Page:
      type: object
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }

    PrincipalRef:
      type: object
      properties:
        type: { type: string, enum: [API_KEY, OAUTH_USER, SYSTEM] }
        id: { type: string }
        displayName: { type: string }

    ApiKeyCreateRequest:
      type: object
      required: [name, role]
      properties:
        name: { type: string, maxLength: 120, minLength: 1 }
        role: { type: string, enum: [PRIVILEGED, STANDARD] }
        expiresAt: { type: string, format: date-time, nullable: true }
    ApiKeyPatchRequest:
      type: object
      properties:
        enabled: { type: boolean }
    ApiKeyView:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        keyPrefix: { type: string }
        role: { type: string, enum: [PRIVILEGED, STANDARD] }
        enabled: { type: boolean }
        bootstrap: { type: boolean }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        lastUsedAt: { type: string, format: date-time, nullable: true }
    ApiKeyCreatedResponse:
      allOf:
        - $ref: '#/components/schemas/ApiKeyView'
        - type: object
          required: [plaintextKey]
          properties:
            plaintextKey: { type: string }
    ApiKeyPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/ApiKeyView' }

    ProfileView:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        preset: { type: string, enum: [BASIC, STANDARD, STRICT, CUSTOM] }
        isDefault: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProfileCreateRequest:
      type: object
      required: [name, preset]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, nullable: true }
        preset: { type: string, enum: [BASIC, STANDARD, STRICT, CUSTOM] }
        policyXml: { type: string, nullable: true, description: "Required when preset=CUSTOM" }
    ProfileUpdateRequest:
      type: object
      properties:
        description: { type: string, nullable: true }
        policyXml: { type: string, nullable: true }
    ProfilePage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/ProfileView' }

    VerificationResponse:
      type: object
      properties:
        verifiedAt: { type: string, format: date-time }
        profileUsed: { type: string }
        overridesApplied: { type: boolean }
        signatureFormat: { type: string }
        indication: { type: string }
        subIndication: { type: string, nullable: true }
        signatureCount: { type: integer }
        reports:
          type: object
          additionalProperties: { type: object }